pyproject.toml
[tool.poetry]
name = "cycl"
version = "0.1.0"
description = "User friendly container management"
authors = ["Oleksii 'LXj' Aleksieiyev <gotletter@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
plumbum = "^1.6.9"
typer = "^0.2.1"
pyyaml = "^5.3.1"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"
isort = "^4.3.21"
flake8 = "^3.8.2"
pre-commit = "^2.4.0"
rope = "^0.17.0"

[tool.poetry.scripts]
cycl = 'cycl.cycl:app'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
